move approvees test to integration, pull up methods to parent
