+++ src/test/java/net/helix/pendulum/AbstractPendulumTest.java
@@ -3,10 +3,13 @@ package net.helix.pendulum;
 import net.helix.pendulum.conf.MainnetConfig;
 import net.helix.pendulum.conf.PendulumConfig;
 import net.helix.pendulum.controllers.TipsViewModel;
+import net.helix.pendulum.controllers.TransactionViewModel;
+import net.helix.pendulum.crypto.SpongeFactory;
 import net.helix.pendulum.event.EventManager;
 import net.helix.pendulum.model.Hash;
 import net.helix.pendulum.model.HashFactory;
 import net.helix.pendulum.model.StateDiff;
+import net.helix.pendulum.model.TransactionHash;
 import net.helix.pendulum.model.persistables.*;
 import net.helix.pendulum.network.Node;
 import net.helix.pendulum.network.Node.RequestQueue;
@@ -36,6 +39,9 @@ import net.helix.pendulum.storage.rocksdb.RocksDBPersistenceProvider;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.rules.TemporaryFolder;
+
+import static net.helix.pendulum.TransactionTestUtils.createTransactionWithHex;
+import static net.helix.pendulum.TransactionTestUtils.getTransactionBytes;
 
 /**
  * Date: 2019-11-18
@@ -160,5 +166,37 @@ public abstract class AbstractPendulumTest {
         tangle.clearColumn(Tag.class);
         tangle.clearColumn(Validator.class);
 
+    }
+
+    protected TransactionViewModel getTxWithBranchAndTrunk() throws Exception {
+        TransactionViewModel tx;
+        TransactionViewModel trunkTx;
+        TransactionViewModel branchTx;
+
+        String hexTx = "0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000c2eb2d5297f4e70f3e40e3d7aa3f5c1d7405264aeb72232d06776605d8b61211000000000000000a0000000000000000000000000000000000000000000000000000000000000000000000005d092fc0000000000000000c000000000000000c5031b48d241283c312c68c777bc4563ddd7cbe1ae6a2c58079e1bf3cfef826790000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000016b6c8a2da00000000000000000000000000000007f00000000000000f40000000000000000000000000000007f00000000000091b0";
+        byte[] bytes = createTransactionWithHex(hexTx).getBytes();
+
+        trunkTx = new TransactionViewModel(bytes, TransactionHash.calculate(SpongeFactory.Mode.S256, bytes));
+        branchTx = new TransactionViewModel(bytes, TransactionHash.calculate(SpongeFactory.Mode.S256, bytes));
+
+        byte[] childTx = getTransactionBytes();
+        System.arraycopy(trunkTx.getHash().bytes(), 0, childTx, TransactionViewModel.TRUNK_TRANSACTION_OFFSET, TransactionViewModel.TRUNK_TRANSACTION_SIZE);
+        System.arraycopy(branchTx.getHash().bytes(), 0, childTx, TransactionViewModel.BRANCH_TRANSACTION_OFFSET, TransactionViewModel.BRANCH_TRANSACTION_SIZE);
+        tx = new TransactionViewModel(childTx, TransactionHash.calculate(SpongeFactory.Mode.S256, childTx));
+
+        trunkTx.store(tangle, snapshotProvider.getInitialSnapshot());
+        branchTx.store(tangle, snapshotProvider.getInitialSnapshot());
+        tx.store(tangle, snapshotProvider.getInitialSnapshot());
+
+        return tx;
+    }
+
+    protected TransactionViewModel getTxWithoutBranchAndTrunk() throws Exception {
+        byte[] bytes = getTransactionBytes();
+        TransactionViewModel tx = new TransactionViewModel(bytes, TransactionHash.calculate(SpongeFactory.Mode.S256, bytes));
+
+        tx.store(tangle, snapshotProvider.getInitialSnapshot());
+
+        return tx;
     }
 }
